# student-dropout-prediction/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Student Dropout Early-Warning System API
  description: |
    Role-Based Backend API for the Student Dropout Early-Warning System.
    Provides endpoints for student risk assessment, mentor intervention tracking,
    cohort reporting, bias & privacy auditing, and mentor management.
    
    ### Roles & Access Levels:
    - **Admin**: Full access across all departments, mentors, audit logs, fairness evaluation, and mentor user management.
    - **Mentor**: Scoped access limited to students assigned to the authenticated mentor. Audit and mentor management endpoints are restricted.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local Development FastAPI Server
security:
  - BearerAuth: []
...
paths:
...
  /api/interventions:
    get:
      tags:
        - Interventions
      summary: Get intervention tracking records
      description: |
        Retrieves students enriched with intervention status, assigned mentor, and notes timeline.
        Mentors only receive their assigned board.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [All, "Not Started", "In Progress", Resolved, Escalated]
        - name: mentor_id
          in: query
          description: Filter by mentor (Admin only)
          schema:
            type: string
        - name: risk_band
          in: query
          schema:
            type: string
            enum: [All, High, Medium, Low]
      responses:
        '200':
          description: List of intervention records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InterventionStudent'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/interventions/{student_id}:
    get:
      tags:
        - Interventions
      summary: Get single student intervention details
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionStudent'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/interventions/{student_id}/status:
    patch:
      tags:
        - Interventions
      summary: Update intervention status
      description: Updates the status of a student follow-up (e.g. Mark as Resolved, Escalate).
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: ["Not Started", "In Progress", Resolved, Escalated]
      responses:
        '200':
          description: Updated intervention record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionStudent'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/interventions/{student_id}/notes:
    post:
      tags:
        - Interventions
      summary: Add mentor progress note
      description: Appends a timestamped follow-up note to the student's intervention timeline.
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  example: Referred student to emergency bursary office.
      responses:
        '201':
          description: Note added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionNote'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/interventions/{student_id}/reassign:
    patch:
      tags:
        - Interventions
      summary: Reassign student to another mentor (Admin Only)
      description: Reassigns the student's lead mentor.
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mentor_name
              properties:
                mentor_name:
                  type: string
                  example: Dr. Priya Nair
      responses:
        '200':
          description: Reassignment successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InterventionStudent'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenAdminOnlyError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ==========================================
  # REPORTS
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Authentication required or invalid/expired session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Access denied for current user or resource not assigned
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenAdminOnlyError:
      description: Restricted endpoint - Admin role required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Student:
      type: object
      properties:
        student_id:
          type: string
          example: STU-1002
        student_name:
          type: string
          example: Marco Silva
        department:
          type: string
          example: Mechanical Engineering
        semester:
          type: integer
          example: 3
        admission_grade:
          type: number
          example: 118.7
        age_at_enrollment:
          type: integer
          example: 22
        scholarship_holder:
          type: boolean
          example: false
        tuition_fees_up_to_date:
          type: boolean
          example: true
        curricular_units_1st_sem_enrolled:
          type: integer
          example: 6
        curricular_units_1st_sem_approved:
          type: integer
          example: 3
        curricular_units_2nd_sem_enrolled:
          type: integer
          example: 6
        curricular_units_2nd_sem_approved:
          type: integer
          example: 2
        curricular_units_failed:
          type: integer
          example: 7
        approval_rate:
          type: number
          example: 0.42
        attendance_percentage:
          type: number
          example: 61
        dropout_probability:
          type: number
          example: 0.81
        risk_category:
          type: string
          enum: [High, Medium, Low]
          example: High
        risk_factors:
          type: array
          items:
            $ref: '#/components/schemas/RiskFactor'
        interventions:
          type: array
          items:
            $ref: '#/components/schemas/InterventionItem'
    RiskFactor:
      type: object
      properties:
        factor:
          type: string
          example: High number of failed curricular units
        tier:
          type: string
          enum: [major, moderate, minor]
          example: major
        direction:
          type: string
          enum: [risk, protective]
          example: risk
    InterventionItem:
      type: object
      properties:
        type:
          type: string
          example: Academic tutoring referral
        date:
          type: string
          format: date
          example: 2026-06-02
        notes:
          type: string
          example: Referred to peer tutoring.
        status:
          type: string
          enum: [Open, "In Progress", Resolved]
          example: In Progress
        mentor_name:
          type: string
          example: Dr. Priya Nair

    # ── Intervention Tracking Schemas
    InterventionStudent:
      allOf:
        - $ref: '#/components/schemas/Student'
        - type: object
          properties:
            intervention_status:
              type: string
              enum: ["Not Started", "In Progress", Resolved, Escalated]
              example: In Progress
            assigned_mentor:
              type: string
              example: Dr. Priya Nair
            last_updated:
              type: string
              format: date
              nullable: true
              example: 2026-07-10
            mentor_notes:
              type: array
              items:
                $ref: '#/components/schemas/InterventionNote'
    InterventionNote:
      type: object
      properties:
        id:
          type: string
          example: n1
        author:
          type: string
          example: Dr. Priya Nair
        timestamp:
          type: string
          format: date-time
          example: 2026-08-17T09:15:00Z
        text:
          type: string
          example: Scheduled follow-up counseling session.

    # ── Dashboard Schemas
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: Invalid credentials or access restricted.